Remove the autogenerated controller scaffolding for ReplicaSchedulingPreferences
As our intention is to move to CRD, it ideally should be ok to use federated
controller implementation similar to how it evolved in federation V1, for
our alpha implementation of the given API type.
In case we don't, we should evaluate the autogenerated scaffolding and
modify (if needed) to suite the federated controller implementation
at a later stage.